zacs/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
Django==4.2.27
django-bootstrap5==24.3
django-cors-headers==4.3.1
djangorestframework==3.15.2
idna==3.11
kombu==5.6.2
Markdown==3.10.1
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
pycparser==3.0
pyspnego==0.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pywinrm==0.4.3
redis==5.2.0
requests==2.32.3
requests_ntlm==1.3.0
six==1.17.0
sqlparse==0.5.0
typing_extensions>=4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2
xmltodict==1.0.2
pyotp
# 新增：日志序列化支持
python-json-logger==2.0.7